Refactor hex grid and tile structure: update namespaces and clean up using directives
